requirements.txt
Django
psycopg2
Pillow
djangorestframework